laser_and_energy_done

+++ SpaceMissile/Assets/Scripts/Model/ShellModel.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class ShellModel : SpaceElement
 {
-    double damage = 5;
+    float damage = 5;
 
-    public double get_damage()
+    public float get_damage()
     {
         return this.damage;
     }
